chore: use supported subgraphs for pool components
