+++ apps/hubble-stats/components/Breadcrumbs/Breadcrumbs.tsx
@@ -8,7 +8,7 @@ import {
 import cx from 'classnames';
 import Link from 'next/link';
 import { usePathname, useRouter } from 'next/navigation';
-import { FC, useMemo } from 'react';
+import { type FC, useMemo, useEffect } from 'react';
 
 import { VANCHORS_MAP } from '../../constants';
 import { BreadcrumbType } from './types';
@@ -48,6 +48,15 @@ const Breadcrumbs: FC = () => {
       },
     ];
   }, [pathname]);
+
+  /**
+   * Pages not reload server data when navigating to a page when go back and forth in the browser
+   * Therefore, need to refresh the page when the user navigates to a page
+   * This keeps all client-side components still remain their states
+   */
+  useEffect(() => {
+    router.refresh();
+  }, [router, pathname]);
 
   return (
     <BreadcrumbsCmp>
