+++ apps/hubble-stats/utils/index.ts
@@ -7,7 +7,6 @@ export { default as getFormattedDataForVolumeChart } from './getFormattedDataFor
 export { default as getRoundedDownNumberWith2Decimals } from './getRoundedDownNumberWith2Decimals';
 export { default as getShortenChainName } from './getShortenChainName';
 export { default as getSortedTypedChainIds } from './getSortedTypedChainIds';
-export { default as getVAnchorMapFromOnChainData } from './getVAnchorMapFromOnChainData';
 export { default as getWrappingFeesPercentageByFungibleToken } from './getWrappingFeesPercentageByFungibleToken';
 export { default as serializeEpochData } from './serializeEpochData';
